+++ relay-server/src/services/buffer/envelope_stack/mod.rs
@@ -1,27 +1,28 @@
-use crate::envelope::Envelope;
 use std::future::Future;
 
+use crate::envelope::Envelope;
+
+pub mod memory;
 pub mod sqlite;
 
 /// A stack-like data structure that holds [`Envelope`]s.
-pub trait EnvelopeStack {
+pub trait EnvelopeStack: Send + std::fmt::Debug {
     /// The error type that is returned when an error is encountered during reading or writing the
     /// [`EnvelopeStack`].
-    type Error;
+    type Error: std::fmt::Debug;
 
     /// Pushes an [`Envelope`] on top of the stack.
-    #[allow(dead_code)]
     fn push(&mut self, envelope: Box<Envelope>) -> impl Future<Output = Result<(), Self::Error>>;
 
     /// Peeks the [`Envelope`] on top of the stack.
-    ///
-    /// If the stack is empty, an error is returned.
-    #[allow(dead_code)]
-    fn peek(&mut self) -> impl Future<Output = Result<&Box<Envelope>, Self::Error>>;
+    fn peek(&mut self) -> impl Future<Output = Result<Option<&Envelope>, Self::Error>>;
 
     /// Pops the [`Envelope`] on top of the stack.
-    ///
-    /// If the stack is empty, an error is returned.
-    #[allow(dead_code)]
-    fn pop(&mut self) -> impl Future<Output = Result<Box<Envelope>, Self::Error>>;
+    fn pop(&mut self) -> impl Future<Output = Result<Option<Box<Envelope>>, Self::Error>>;
+}
+
+pub trait StackProvider: std::fmt::Debug {
+    type Stack: EnvelopeStack;
+
+    fn create_stack(&self, envelope: Box<Envelope>) -> Self::Stack;
 }