+++ relay-server/src/services/buffer/stack_provider/memory.rs
@@ -1,0 +1,14 @@
+use crate::services::buffer::envelope_stack::memory::MemoryEnvelopeStack;
+use crate::services::buffer::envelope_stack::StackProvider;
+use crate::Envelope;
+
+#[derive(Debug)]
+pub struct MemoryStackProvider;
+
+impl StackProvider for MemoryStackProvider {
+    type Stack = MemoryEnvelopeStack;
+
+    fn create_stack(&self, envelope: Box<Envelope>) -> Self::Stack {
+        MemoryEnvelopeStack::new(envelope)
+    }
+}